Change input variable to output_directory
This is now the directory in the current workspace bucket where the
backup files will be saved. Because it can be derive from the
workspace namespace and name, there will likley be less user error
when running this workflow from different workspaces. Note that this
means that you cannot use this workflow to write the files to a
different bucket than the workspace where they exist.